Updates to hospital tree R file
The hospital trees just predict that people will be hospitalized if
they have an ulcer or congestive heart failure and are very sensitive
to the seed. It looks like they will need to be bagged or boosted, but
may not really capture what we consider the “true health” that we want
to predict.